chore!: spigot -> bukkit, use version catalogs, implement UserChatEvent#setMessage
