+++ annotations/src/main/java/dev/hypera/chameleon/annotations/Plugin.java
@@ -23,9 +23,9 @@
 package dev.hypera.chameleon.annotations;
 
 import dev.hypera.chameleon.annotations.processing.generation.Generator;
+import dev.hypera.chameleon.annotations.processing.generation.impl.bukkit.BukkitGenerator;
 import dev.hypera.chameleon.annotations.processing.generation.impl.bungeecord.BungeeCordGenerator;
 import dev.hypera.chameleon.annotations.processing.generation.impl.minestom.MinestomGenerator;
-import dev.hypera.chameleon.annotations.processing.generation.impl.spigot.SpigotGenerator;
 import dev.hypera.chameleon.annotations.processing.generation.impl.velocity.VelocityGenerator;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -104,7 +104,7 @@ public @interface Plugin {
      */
     enum Platform {
 
-        BUNGEECORD(BungeeCordGenerator.class), MINESTOM(MinestomGenerator.class), SPIGOT(SpigotGenerator.class), VELOCITY(VelocityGenerator.class);
+        BUNGEECORD(BungeeCordGenerator.class), MINESTOM(MinestomGenerator.class), BUKKIT(BukkitGenerator.class), VELOCITY(VelocityGenerator.class);
 
         private final @NotNull Class<? extends Generator> generator;
