+++ core/build.gradle.kts
@@ -31,12 +31,12 @@ val tokens = mapOf(
 )
 
 dependencies {
-    api("net.kyori:adventure-api:4.11.0")
-    api("net.kyori:adventure-text-serializer-legacy:4.11.0")
-    api("net.kyori:adventure-text-serializer-gson:4.11.0")
-    api("net.kyori:adventure-platform-api:4.1.1")
-    compileOnly("org.slf4j:slf4j-api:1.7.36")
-    compileOnlyApi("org.jetbrains:annotations:23.0.0")
+    api(libs.adventure.api)
+    api(libs.adventure.textSerializer.legacy)
+    api(libs.adventure.textSerializer.gson)
+    api(libs.adventure.platform.api)
+    compileOnly(libs.slf4j)
+    compileOnlyApi(libs.annotations)
 }
 
 val sourcesForRelease = task<Copy>("sourcesForRelease") {
