+++ core/src/main/java/dev/hypera/chameleon/core/data/PluginData.java
@@ -114,7 +114,7 @@ public class PluginData {
      * Platform.
      */
     public enum Platform {
-        BUNGEECORD, MINESTOM, SPIGOT, VELOCITY
+        BUNGEECORD, MINESTOM, BUKKIT, VELOCITY
     }
 
 }
