+++ platform-bungeecord/build.gradle.kts
@@ -32,8 +32,8 @@ repositories {
 }
 
 dependencies {
-    api(project(":core"))
-    compileOnly("net.md-5:bungeecord-api:1.19-R0.1-SNAPSHOT")
-    implementation("net.kyori:adventure-platform-bungeecord:4.1.1")
-    compileOnlyApi("org.jetbrains:annotations:23.0.0")
+    api(project(":chameleon-core"))
+    compileOnly(libs.platform.bungeecord)
+    implementation(libs.adventure.platform.bungeecord)
+    compileOnlyApi(libs.annotations)
 }
