--- platform-spigot/src/main/java/dev/hypera/chameleon/platforms/spigot/user/SpigotConsoleUser.java
@@ -1,65 +1,0 @@
-/*
- * Chameleon Framework - Cross-platform Minecraft plugin framework
- *  Copyright (c) 2021-present The Chameleon Framework Authors.
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in all
- *  copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- *  SOFTWARE.
- */
-package dev.hypera.chameleon.platforms.spigot.user;
-
-import dev.hypera.chameleon.core.Chameleon;
-import dev.hypera.chameleon.core.adventure.AbstractAudience;
-import dev.hypera.chameleon.core.users.ChatUser;
-import org.bukkit.Bukkit;
-import org.jetbrains.annotations.ApiStatus.Internal;
-import org.jetbrains.annotations.NotNull;
-
-/**
- * Spigot console {@link ChatUser} implementation.
- */
-@Internal
-public class SpigotConsoleUser extends AbstractAudience implements ChatUser {
-
-    /**
-     * {@link SpigotConsoleUser} constructor.
-     *
-     * @param chameleon {@link Chameleon} instance.
-     */
-    @Internal
-    public SpigotConsoleUser(@NotNull Chameleon chameleon) {
-        super(chameleon.getAdventure().console());
-    }
-
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public @NotNull String getName() {
-        return Bukkit.getConsoleSender().getName();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean hasPermission(@NotNull String permission) {
-        return true;
-    }
-
-}
