--- platform-spigot/src/main/java/dev/hypera/chameleon/platforms/spigot/user/SpigotUsers.java
@@ -1,60 +1,0 @@
-/*
- * Chameleon Framework - Cross-platform Minecraft plugin framework
- *  Copyright (c) 2021-present The Chameleon Framework Authors.
- *
- *  Permission is hereby granted, free of charge, to any person obtaining a copy
- *  of this software and associated documentation files (the "Software"), to deal
- *  in the Software without restriction, including without limitation the rights
- *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *  copies of the Software, and to permit persons to whom the Software is
- *  furnished to do so, subject to the following conditions:
- *
- *  The above copyright notice and this permission notice shall be included in all
- *  copies or substantial portions of the Software.
- *
- *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- *  SOFTWARE.
- */
-package dev.hypera.chameleon.platforms.spigot.user;
-
-import dev.hypera.chameleon.core.Chameleon;
-import dev.hypera.chameleon.core.users.ChatUser;
-import dev.hypera.chameleon.platforms.spigot.SpigotChameleon;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-import org.jetbrains.annotations.ApiStatus.Internal;
-import org.jetbrains.annotations.NotNull;
-
-/**
- * Spigot {@link dev.hypera.chameleon.core.users.User} utilities.
- */
-@Internal
-public final class SpigotUsers {
-
-    @Internal
-    private SpigotUsers() {
-
-    }
-
-    /**
-     * Wrap provided {@link CommandSender}.
-     *
-     * @param chameleon {@link Chameleon} instance.
-     * @param sender    {@link CommandSender} to wrap.
-     *
-     * @return {@link ChatUser}.
-     */
-    public static @NotNull ChatUser wrap(@NotNull Chameleon chameleon, @NotNull CommandSender sender) {
-        if (sender instanceof Player) {
-            return new SpigotUser((SpigotChameleon) chameleon, (Player) sender);
-        } else {
-            return new SpigotConsoleUser(chameleon);
-        }
-    }
-
-}
