+++ platform-velocity/build.gradle.kts
@@ -29,7 +29,7 @@ repositories {
 }
 
 dependencies {
-    api(project(":core"))
-    compileOnly("com.velocitypowered:velocity-api:3.1.0")
-    compileOnlyApi("org.jetbrains:annotations:23.0.0")
+    api(project(":chameleon-core"))
+    compileOnly(libs.platform.velocity)
+    compileOnlyApi(libs.annotations)
 }
